% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_config.R
\name{gen_config}
\alias{gen_config}
\title{Generate a config from a site with a Tableau Dashboard.}
\usage{
gen_config(site, which = 1)
}
\arguments{
\item{site}{The url of a site that has an embedded dashboard.}

\item{which}{If the site contains multiple dashboards, which one to use.}
}
\value{
A list of parameters to be used by \code{scrape_tableau()}.
If the site has more than one dashboard, only the first is returned, with a warning.
}
\description{
Generate a config from a site with a Tableau Dashboard.
}
\examples{
\dontrun{
config <- gen_config(
  "https://www.phila.gov/programs/coronavirus-disease-2019-covid-19/vaccines/data/"
)
dfs <- scrape_tableau(config)
}

}
