% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{scrape_tableau}
\alias{scrape_tableau}
\title{Scrape a tableau dashboard.}
\usage{
scrape_tableau(config)
}
\arguments{
\item{config}{A list with fields \code{host_url}, \code{path}, and \code{req_url}.
Can be generated by \code{gen_config()}.}
}
\value{
A list of dataframes behind the dashboard.
}
\description{
Scrape a tableau dashboard.
}
\examples{
\dontrun{
config <- gen_config(
  "https://www.phila.gov/programs/coronavirus-disease-2019-covid-19/vaccines/data/"
)
dfs <- scrape_tableau(config)
}

}
